Fix Linux compilation and production
git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@311 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ library/ctlibstructureaddon/ct_itemdrawable/ct_grid4d_dense.hpp
@@ -142,7 +142,7 @@ CT_Grid4D_Dense<DataT>::~CT_Grid4D_Dense()
 template< typename DataT>
 void CT_Grid4D_Dense<DataT>::initGridWithValue(const DataT val)
 {
-    for (int  i = 0 ; i < this->nCells() ; i++)
+    for (size_t  i = 0 ; i < this->nCells() ; i++)
     {
         _data(i) = val;
     }
